lwan-response.c: Fix integer conversion warnings

+++ common/strbuf.c
@@ -227,7 +227,7 @@ strbuf_append_printf(strbuf_t *s, const char *fmt, ...)
     return could_printf;
 }
 
-ALWAYS_INLINE int
+ALWAYS_INLINE size_t
 strbuf_get_length(strbuf_t *s)
 {
     return s->len.buffer;
